renamed to layout_blocks
