+++ build.gradle
@@ -12,8 +12,8 @@ apply plugin: 'net.minecraftforge.gradle.forge'
 
 
 version = "1.0"
-group = "com.yourname.modid" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
-archivesBaseName = "testMod"
+group = "com.schoggomilch.layout_blocks" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+archivesBaseName = "LayoutBlocks"
 
 sourceCompatibility = targetCompatibility = '1.8' // Need this here so eclipse task generates correctly.
 compileJava {
