+++ src/main/java/com/schoggomilch/layout_blocks/block/LayoutBlocksMaterials.java
@@ -1,0 +1,21 @@
+package com.schoggomilch.layout_blocks.block;
+
+import net.minecraft.block.material.MapColor;
+import net.minecraft.block.material.Material;
+
+public class LayoutBlocksMaterials {
+        public static final Material NonsolidMat = new Material(MapColor.GRAY){
+        public boolean isSolid()
+        {
+            return false;
+        }
+        public boolean blocksLight()
+        {
+            return false;
+        }
+        public boolean blocksMovement()
+        {
+            return false;
+        }
+        };
+}
