+++ src/main/java/com/schoggomilch/layout_blocks/gui/RenderGuiHandler.java
@@ -1,0 +1,33 @@
+package com.schoggomilch.layout_blocks.gui;
+
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.inventory.Container;
+import net.minecraft.world.World;
+import net.minecraftforge.fml.common.network.IGuiHandler;
+
+
+public class RenderGuiHandler implements IGuiHandler {
+
+    public static final int PLACER =0;
+
+    @Override
+    public Container getServerGuiElement(int ID, EntityPlayer player, World world, int x, int y, int z) {
+        switch (ID) {
+            default:
+                return null;
+        }
+    }
+
+
+    @Override
+    public Object getClientGuiElement(int ID, EntityPlayer player, World world, int x, int y, int z) {
+        switch (ID) {
+            case 0:
+                return new PlacerGui(player, player.getHeldItemMainhand());
+            default:
+                return null;
+        }
+    }
+
+
+}
