+++ src/main/java/com/schoggomilch/layout_blocks/init/ModBlocks.java
@@ -1,0 +1,9 @@
+package com.schoggomilch.layout_blocks.init;
+
+import com.schoggomilch.layout_blocks.block.BlockNonsolidLayoutBlock;
+import com.schoggomilch.layout_blocks.block.BlockSolidLayoutBlock;
+
+public class ModBlocks {
+    public static BlockSolidLayoutBlock blockSolidLayoutBlock = new BlockSolidLayoutBlock();
+    public static BlockNonsolidLayoutBlock blockNonsolidLayoutBlock = new BlockNonsolidLayoutBlock();
+}
