+++ src/main/java/com/schoggomilch/layout_blocks/proxy/CommonProxy.java
@@ -1,0 +1,32 @@
+package com.schoggomilch.layout_blocks.proxy;
+
+
+import net.minecraft.item.Item;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+@Mod.EventBusSubscriber
+public class CommonProxy {
+
+
+    public void preInit(FMLPreInitializationEvent e) {
+    }
+
+    public void init(FMLInitializationEvent e) {
+    }
+
+    public void postInit(FMLPostInitializationEvent e) {
+    }
+
+
+
+    public void registerItemRenderer(Item i, int meta, String id){
+    }
+
+
+
+
+
+}
