--- src/main/java/com/schoggomilch/testmod/init/ModBlocks.java
@@ -1,14 +1,0 @@
-package com.schoggomilch.testmod.init;
-
-import com.schoggomilch.testmod.block.BlockNonsolidLayoutBlock;
-import com.schoggomilch.testmod.block.BlockSolidLayoutBlock;
-import net.minecraft.block.material.Material;
-import net.minecraftforge.common.util.EnumHelper;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ModBlocks {
-    public static BlockSolidLayoutBlock blockSolidLayoutBlock = new BlockSolidLayoutBlock();
-    public static BlockNonsolidLayoutBlock blockNonsolidLayoutBlock = new BlockNonsolidLayoutBlock();
-}
