--- src/main/java/com/schoggomilch/testmod/proxy/CommonProxy.java
@@ -1,40 +1,0 @@
-package com.schoggomilch.testmod.proxy;
-
-
-import com.schoggomilch.testmod.block.BlockSolidLayoutBlock;
-import com.schoggomilch.testmod.item.ItemPlacer;
-import net.minecraft.block.Block;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemBlock;
-import net.minecraftforge.client.event.ModelRegistryEvent;
-import net.minecraftforge.event.RegistryEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.common.network.NetworkRegistry;
-
-@Mod.EventBusSubscriber
-public class CommonProxy {
-
-
-    public void preInit(FMLPreInitializationEvent e) {
-    }
-
-    public void init(FMLInitializationEvent e) {
-    }
-
-    public void postInit(FMLPostInitializationEvent e) {
-    }
-
-
-
-    public void registerItemRenderer(Item i, int meta, String id){
-    }
-
-
-
-
-
-}
